+++ sdks/python/apache_beam/runners/dataflow/dataflow_runner.py
@@ -136,6 +136,10 @@ class DataflowRunner(PipelineRunner):
   def is_fnapi_compatible(self):
     return False
 
+  def apply(self, transform, input, options):
+    self._maybe_add_unified_worker_missing_options(options)
+    return super(DataflowRunner, self).apply(transform, input, options)
+
   def _get_unique_step_name(self):
     self._unique_step_id += 1
     return 's%s' % self._unique_step_id
@@ -434,6 +438,8 @@ class DataflowRunner(PipelineRunner):
           'Google Cloud Dataflow runner not available, '
           'please install apache_beam[gcp]')
 
+    self._maybe_add_unified_worker_missing_options(options)
+
     # Convert all side inputs into a form acceptable to Dataflow.
     if apiclient._use_fnapi(options):
       pipeline.visit(
@@ -565,6 +571,17 @@ class DataflowRunner(PipelineRunner):
     self._metrics = DataflowMetrics(self.dataflow_client, result, self.job)
     result.metric_results = self._metrics
     return result
+
+  def _maybe_add_unified_worker_missing_options(self, options):
+    # set default beam_fn_api and use_beam_bq_sink experiment if use unified
+    # worker experiment flag exists, no-op otherwise.
+    debug_options = options.view_as(DebugOptions)
+    from apache_beam.runners.dataflow.internal import apiclient
+    if apiclient._use_unified_worker(options):
+      if not debug_options.lookup_experiment('beam_fn_api'):
+        debug_options.add_experiment('beam_fn_api')
+      if not debug_options.lookup_experiment('use_beam_bq_sink'):
+        debug_options.add_experiment('use_beam_bq_sink')
 
   def _get_typehint_based_encoding(self, typehint, window_coder):
     """Returns an encoding based on a typehint object."""
@@ -780,7 +797,9 @@ class DataflowRunner(PipelineRunner):
 
     # TODO(BEAM-6928): Remove this function for release 2.14.0.
     experiments = options.view_as(DebugOptions).experiments or []
-    if (not isinstance(transform, beam.io.WriteToBigQuery) or
+    from apache_beam.runners.dataflow.internal import apiclient
+    use_fnapi = apiclient._use_fnapi(options)
+    if (not isinstance(transform, beam.io.WriteToBigQuery) or use_fnapi or
         'use_beam_bq_sink' in experiments):
       return self.apply_PTransform(transform, pcoll, options)
     if transform.schema == beam.io.gcp.bigquery.SCHEMA_AUTODETECT: